pprocess.py: Revert back to python2
There was an issue in 0c23ed97b that we missed in our CI testing and
the graphs are not being produced as expected. So, for now, we revert
to python2 and add one error check. We will reopen issue #8.

We also need to add the python2 package install to the smoke-test
workflow so that it does not fail.

Unfixes #8.